Split module contents
Split `Coordinate` and `Line` classes into two modules since `Line` uses the `Coordinate` class and thus, should not be at the same module level.
